Added some webserver routes
Possibilities:
sequence/<sequence file name>(/<repeatcount>) to execute an arbitrary
sequence.

behavior/<behaviortype> to swtich behaviortype

control/walk/<x>/<y> (logic not implemented yet) to make the spider
walk. X/Y are the values of the analog control stick on the controller.

control/reset to reset the spider to it's initial position

control/stop (logic not implemented yet) to stop the current sequence or
stop walking.